ipc: add compound ipc support
1. add offset in mailbox_validate, and we can read more than one message
from mailbox. 2. use a global variable "reply_offset" to indicate the
offset of writtig to the mailbox. So that we can send more than one repllies
to host. 3. call ipc_glb_decompound_message when we get a COMPOUND type
ipc message. It will decompound the message and handle each message.

Signed-off-by: Bard liao <bard.liao@intel.com>